refactor: DocumentPage, EditorPage 파일 분리
