+++ src/api.js
@@ -10,7 +10,6 @@ export const request = async (url, option = {}) => {
       },
     });
     if (res.ok) {
-      alert("api 정상 호출");
       return res.json();
     }
     throw new Error("API 호출 오류");
